Simplify NewClient to close SSLContext, add memory leak checks in tests
Simplify NewClient to delegate the creation and cleanup of SSLContext to
the Dialer and Conn if there is no SSLContext passed to it. In this
case, the Dialer will create a new context every connection, which is
slower but more memory safe since the caller doesn't need to cleanup
SSLContext.

Add memory leak checks in all the tests so we can ensure we cleanup all
heap-allocated C objects.