fipstls: Add ProviderInfo and log it in dialer
Add ProviderInfo which will have its result cached. Same for Version.
This way we don't need to make unnecessary calls into C since the
library should only be loaded once per program execution.

Log the provider info in the dialer so we can verify we are loading the
expected libs.